base typescript conversion
