+++ frontend/jest.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  preset: '@vue/cli-plugin-unit-jest'
+  preset: '@vue/cli-plugin-unit-jest/presets/typescript-and-babel'
 }
