+++ frontend/package.json.md5
@@ -1,1 +1,1 @@
-2476438886921bbdf9ab4cdcff3927bf
+c90c7ccb4d5908ffe2424d0fc750e54c
