+++ frontend/src/icons.ts
@@ -1,0 +1,75 @@
+import { library } from "@fortawesome/fontawesome-svg-core";
+
+import {
+  faCheck,
+  faCheckCircle,
+  faInfoCircle,
+  faExclamationTriangle,
+  faExclamationCircle,
+  faArrowUp,
+  faAngleRight,
+  faAngleLeft,
+  faAngleDown,
+  faEye,
+  faEyeSlash,
+  faCaretDown,
+  faCaretUp,
+  faUpload,
+  faUser,
+  faHome,
+  faTree,
+  faTasks,
+  faProjectDiagram,
+  faLeaf,
+  faAddressBook,
+  faVolumeUp,
+  faCog,
+  faLink,
+  faCut,
+  faHeadphones,
+  faFolder,
+  faPencilAlt,
+  faPlus,
+  faRedo,
+  faReply,
+  faSearch,
+} from "@fortawesome/free-solid-svg-icons";
+
+library.add(
+  faCheck,
+  faCheckCircle,
+  faInfoCircle,
+  faExclamationTriangle,
+  faExclamationCircle,
+  faArrowUp,
+  faAngleRight,
+  faAngleLeft,
+  faAngleDown,
+  faEye,
+  faEyeSlash,
+  faCaretDown,
+  faCaretUp,
+  faUpload,
+  faHome,
+  faUser,
+  faTree,
+  faTasks,
+  faLeaf,
+  faAddressBook,
+  faProjectDiagram,
+  faVolumeUp,
+  faCog,
+  faLink,
+  faCut,
+  faHeadphones,
+  faFolder,
+  faPencilAlt,
+  faPlus,
+  faReply,
+  faRedo,
+  faSearch
+);
+
+import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
+
+export default FontAwesomeIcon;
