+++ frontend/src/views/Home.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="home">
+    <img alt="Vue logo" src="../assets/logo.png" />
+    <HelloWorld msg="Welcome to Your Vue.js + TypeScript App" />
+  </div>
+</template>
+
+<script lang="ts">
+import Vue from "vue";
+import HelloWorld from "../components/HelloWorld.vue"; // @ is an alias to /src
+
+export default Vue.extend({
+  name: "Home",
+  components: {
+    HelloWorld,
+  },
+});
+</script>
